mktables: Fix up version compare
This is a feature that is used to compare 2 different Unicode versions
for changes as opposed to additions.  It does this by removing the
newly-added code points in the later version.  One can then diff the
directory structure generated with the older version to see what
changed.

Prior to This commit, it assumed all versions were single digit for the
major number.  This no longer works for Unicode 10.  It also more
completely adds back to the unallocated pool, blocks that didn't exist
in the earlier version.  This gives better diff results.